fix(extractor+server): torch_dtype→dtype, GPU semaphore for concurrent extraction
extractor.py:
- Change torch_dtype= to dtype= in Qwen2VLForConditionalGeneration.from_pretrained()
  (torch_dtype is deprecated in transformers 4.57.x)

server.py:
- Add _gpu_semaphore = threading.Semaphore(1) to serialize GPU extraction jobs
- Wrap run_full_extraction() call with semaphore to prevent VRAM contention
  when concurrent requests arrive

+++ ai-ml/extractor.py
@@ -253,7 +253,7 @@ def _load_qwen() -> None:
         logger.info("[qwen] Loading model weights — dtype=%s  attention=%s", _dtype, _attn_label)
         print(f"[qwen] Loading model weights — dtype={_dtype}  attention={_attn_label}", flush=True)
 
-        _load_kwargs: dict = dict(config=config, torch_dtype=_dtype, device_map="auto")
+        _load_kwargs: dict = dict(config=config, dtype=_dtype, device_map="auto")
         if _attn:
             _load_kwargs["attn_implementation"] = _attn
